Refactor package structure for OpenAPI generation
- Move OpenAPI-related classes under `io.github.nomisrev.openapi.generation` package for better organization.
- Add missing imports and adjust indentation in modified files.
- Update references across the project to reflect the new package structure.
