+++ generation/src/main/kotlin/io/github/nomisrev/openapi/generation/OpenAPIContext.kt
@@ -1,4 +1,4 @@
-package io.github.nomisrev.openapi
+package io.github.nomisrev.openapi.generation
 
 import com.squareup.kotlinpoet.FileSpec
 import java.util.concurrent.atomic.AtomicReference
